requirements.txt
asgiref==3.7.2
confluent-kafka==2.3.0
Django==4.2.5
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==19.6.2
mysqlclient==2.2.0
Pillow==10.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.41.3
